pyproject.toml
[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[project]
name = "plane-cli"
version = "0.2.4"
description = "Plane.so CLI"
requires-python = ">=3.10"
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "typer>=0.12.0",
    "plane-sdk>=0.2.0",
    "tomlkit>=0.13.0",
]

[project.scripts]
plane = "plane_cli.main:app"

[tool.uv.build-backend]
module-name = "plane_cli"
module-root = ""

[dependency-groups]
dev = ["ruff>=0.9.0"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
